Feature/controller

See merge request s05-webmobile1-sub3/S05P13B202!13

+++ backend/src/main/java/com/ssafy/niceage/Domain/Senior_Citizen_Center/Senior_Citizen_Center.java
@@ -14,7 +14,9 @@ import lombok.Data;
 public class Senior_Citizen_Center {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column (name = "senior_id")
-	private int seniorId;
+
+	private Long seniorId;
+
 	
 	@OneToMany(mappedBy = "senior", cascade = CascadeType.ALL, orphanRemoval =true)
 	List<Enter> enters = new ArrayList<Enter>();
@@ -23,5 +25,7 @@ public class Senior_Citizen_Center {
     private String seniorName;
 	
 	@Column (name = "senior_address", nullable = false, length = 200)
+
     private String seniorAddress;
+
 }
